requirements.txt
beautifulsoup4==4.13.4
openai==1.76.0
pandas==2.2.3
pyodbc==5.2.0
python-dotenv==1.1.0
pytz==2024.1
Scrapy==2.11.0
selenium==4.31.0
